refactor: wrap server code in class
