+++ src/huemon/commands/command_handler.py
@@ -3,16 +3,15 @@
 # This source code is licensed under the MPL-2.0 license found in the
 # LICENSE file in the root directory of this source tree.
 
+import os
 from functools import reduce
 
 from huemon.api.api_factory import create_api
 from huemon.api.api_interface import ApiInterface
 from huemon.commands.hue_command_interface import HueCommand
-from huemon.commands_internal.agent_command import AgentCommand
-from huemon.commands_internal.install_available_command import InstallAvailableCommand
 from huemon.infrastructure.logger_factory import create_logger
 from huemon.plugin_loader import load_plugins
-from huemon.util import exit_fail
+from huemon.util import create_local_path, exit_fail
 
 LOG = create_logger()
 
@@ -34,10 +33,11 @@ def __load_command_plugins(config: dict, command_plugins_path: str):
 
 
 def __load_plugins_and_hardwired_handlers(config: dict, command_plugins_path: str):
+    hardwired_commands_path = create_local_path(os.path.join("commands_internal"))
+
     return {
         **__load_command_plugins(config, command_plugins_path),
-        InstallAvailableCommand.name(): InstallAvailableCommand(config),
-        AgentCommand.name(): AgentCommand(config),
+        **__load_command_plugins(config, hardwired_commands_path),
     }
